+++ client/gld_dashboard/src/apps/earn/components/decrease-stake/components/details-dissolving-instantly/hooks/useDissolvingInstantly.tsx
@@ -1,6 +1,6 @@
 import { useQueryClient, useMutation } from "@tanstack/react-query";
 import { Actor, Agent, HttpAgent } from "@dfinity/agent";
-import { idlFactory as idlFactoryStake } from "@services/gldt_stake/interfaces/idlFactory";
+import { idlFactory as idlFactoryStake } from "@services/gldt_stake/idlFactory";
 import dissolve_instantly from "@services/gldt_stake/dissolve_instantly";
 
 const useDissolvingInstantly = (