+++ client/gld_dashboard/src/apps/earn/components/increase-stake/components/details/index.tsx
@@ -20,18 +20,13 @@ const DetailsDialogContent = () => {
 
   const handleAddStake = () => {
     if (!stakeState.user_balance_gldt) return;
-    const stakeAmountE8s = BigInt(
+    const stakeAmount = BigInt(
       Math.round(
         Number(stakeState.stake_amount) *
           10 ** stakeState.user_balance_gldt.decimals
       )
     );
-    const isAmountMaxBalance =
-      stakeAmountE8s + stakeState.user_balance_gldt.fee_e8s >=
-      stakeState.user_balance_gldt.balance_e8s;
-    const amount = isAmountMaxBalance
-      ? stakeAmountE8s
-      : stakeAmountE8s + stakeState.user_balance_gldt.fee_e8s;
+    const amount = stakeAmount + 1n * stakeState.user_balance_gldt.fee_e8s;
     addStake.mutate({
       amount,
     });