+++ client/gld_dashboard/src/apps/earn/hooks/useFetchStakeAPY.tsx
@@ -4,7 +4,7 @@ import {
   UseQueryOptions,
 } from "@tanstack/react-query";
 import { Actor, Agent, HttpAgent } from "@dfinity/agent";
-import { idlFactory } from "@services/gldt_stake/interfaces/idlFactory";
+import { idlFactory } from "@services/gldt_stake/idlFactory";
 import get_apy_timeseries from "@services/gldt_stake/get_apy_timeseries";
 
 const useFetchStakeAPY = (